+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/common/BaseRoute.kt
@@ -1,0 +1,5 @@
+package com.mashup.gabbangzip.sharedalbum.presentation.common
+
+interface BaseRoute {
+    val route: String
+}
