+++ presentation/src/main/java/com/mashup/gabbangzip/sharedalbum/presentation/groupMake/navigation/GroupMakeNavigation.kt
@@ -4,7 +4,7 @@ import androidx.activity.compose.BackHandler
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.mashup.gabbangzip.sharedalbum.presentation.common.GroupMakeRoute
+import com.mashup.gabbangzip.sharedalbum.presentation.common.BaseRoute
 import com.mashup.gabbangzip.sharedalbum.presentation.groupMake.GroupMakeScreen
 
 fun NavController.navigateGroupMake() {
@@ -20,4 +20,8 @@ fun NavGraphBuilder.groupMakeNavGraph(
             onBackPressed()
         }
     }
+}
+
+data object GroupMakeRoute : BaseRoute {
+    override val route: String = "groupMake"
 }
